% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tgam.R
\name{tgam_cols}
\alias{tgam_cols}
\title{Function to extract The Globe and Mail's colors as hex codes}
\usage{
tgam_cols(...)
}
\arguments{
\item{...}{Character names of tgam_colors. Colour names include 'burgundy',
'sky', 'salmon', 'gold', 'gray', 'alt1', 'alt2', 'alt3', 'alt4', 'alt5'.}
}
\value{
A character vector of one or more colours in The Globe's palette
}
\description{
Function to extract The Globe and Mail's colors as hex codes
}
\examples{
tgam_cols()
tgam_cols('burgundy')

}
