% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tgam.R
\name{tgam_pal}
\alias{tgam_pal}
\title{Return function to interpolate a Globe and Mail color palette}
\usage{
tgam_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in tgam_palettes. Currently, one of 'main' or 'alternate'.}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{
A character vector of interpolated colour values
}
\description{
Return function to interpolate a Globe and Mail color palette
}
\examples{
tgam_pal()(3)
tgam_pal(palette = 'alternate')(3)

}
