% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tgam.R
\name{theme_tgam}
\alias{theme_tgam}
\title{ggplot Globe and Mail theme}
\usage{
theme_tgam(
  size = 12,
  family = "GMsanC-Regular",
  title = "GMsanC-Bold",
  position = "center"
)
}
\arguments{
\item{size}{Base font size}

\item{family}{Font family}

\item{title}{Font to use for title text}

\item{position}{How should headings be positioned? Options are 'center' or 'left'. (Default: 'center',)}
}
\value{
A ggplot2 theme object, to be passed to a ggplot2 object
}
\description{
Theme function to be passed as an argument during ggplot2 charting.
}
